plantsense_voice/pyproject.toml
[project]
name = "plantsense-voice"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "httpx>=0.27.0",
]
